fix: post-gym roll party-full routing and wrong message (v2 done)
targetSlot was read as u16 in ScrCmd_ExecuteGymRoll so -1 from the
script arrived as 0xFFFF, making the party-or-box branch unreachable.
DraftManager_AddPokemon now uses Party_AddPokemon's return value to
route to the box when full and sets sentToPC for the correct message.

+++ src/scrcmd_party.c
@@ -847,7 +847,7 @@ BOOL ScrCmd_ExecuteGymRoll(ScriptContext *ctx)
         return FALSE;
     }
 
-    u16 targetSlot = ScriptContext_GetVar(ctx);
+    s16 targetSlot = (s16)ScriptContext_GetVar(ctx);
     u16 level = ScriptContext_GetVar(ctx);
     int metLocation = MapHeader_GetMapLabelTextID(ctx->fieldSystem->location->mapId);
